feat: file based configuration reloading support, relocate http
This change moves the http listener utility for the *http.Server
out of the api package. In addition we implement http.Handler on
the *api.API object so we may wrap it with the recently committed
*reloader.AtomicHandler. I've also removed the sync.WaitGroup that
previously coordinated shutdown with main.go. A single environment
based configuration is still used at startup, but these changes
set us up to add config reloads in an upcoming commit.

Note:
I added a sync.WaitGroup to serve, this means the maximum shutdown
time is now time.Minute*2. More intrusive options to fix this
exist if the team thinks it's important.